requirements.txt
requests~=2.27.1

opencv-python~=4.5.5.62
Pillow~=8.4.0
numpy~=1.19.5